Merged PR 8100: Better integration with qiskit APIs
This PR improves the `AzureQuantumJob` such that the `results()` method returns data in the right format such that `get_counts` works correctly and can be used in methods like `plot_histogram`.
Similarly, it trivially implements `queue_position` such that we can use `job_monitor` to check and print the status of a Job.