Add new azure-quantum package release pipeline (#526)
This PR creates a new release pipeline for azure-quantum SDK that is separate from the existing qdk.release pipeline.

Currently, this pipeline allows us to release using semver and iterate the patch version from the pipeline itself. We can then use it to publish packages to PyPi and create a draft GitHub release.

Other items addressed in the PR include updating the repo to the new URL and preparing for the v1 of qsharp package by locking to <1.0.